Feature: Read-only access

* Fix API access rights for readonly

* Update no-access page

* Edit required role of edit and new pages

* Update event view

* Merge branch 'dev' into feature/read-only


Approved-by: Johan Sjöberg

+++ src/pages/api/equipment/[id].ts
@@ -1,5 +1,6 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 import {
+    respondWithAccessDeniedResponse,
     respondWithCustomErrorMessage,
     respondWithEntityNotFoundResponse,
     respondWithInvalidDataResponse,
@@ -10,10 +11,11 @@ import {
     updateEquipment,
     validateEquipmentObjectionModel,
 } from '../../../lib/db-access';
-import { withSessionContext } from '../../../lib/sessionContext';
+import { SessionContext, withSessionContext } from '../../../lib/sessionContext';
+import { Role } from '../../../models/enums/Role';
 
 const handler = withSessionContext(
-    async (req: NextApiRequest, res: NextApiResponse): Promise<Promise<void> | void> => {
+    async (req: NextApiRequest, res: NextApiResponse, context: SessionContext): Promise<Promise<void> | void> => {
         if (isNaN(Number(req.query.id))) {
             respondWithEntityNotFoundResponse(res);
             return;
@@ -26,11 +28,21 @@ const handler = withSessionContext(
                     .catch((error) => respondWithCustomErrorMessage(res, error.message));
 
             case 'DELETE':
+                if (context.currentUser.role == Role.READONLY) {
+                    respondWithAccessDeniedResponse(res);
+                    return;
+                }
+
                 return deleteEquipment(Number(req.query.id))
                     .then((result) => res.status(200).json(result))
                     .catch((error) => respondWithCustomErrorMessage(res, error.message));
 
             case 'PUT':
+                if (context.currentUser.role == Role.READONLY) {
+                    respondWithAccessDeniedResponse(res);
+                    return;
+                }
+                
                 if (!validateEquipmentObjectionModel(req.body.equipment)) {
                     respondWithInvalidDataResponse(res);
                     return;
